housekeeping

+++ src/core/cpu/CPU.h
@@ -2,7 +2,6 @@
 
 #include <vector>
 #include <cstdint>
-#include <string>
 
 class IBus;
 class CPU {
@@ -47,7 +46,7 @@ class CPU {
         uint16_t rel_addr;
         uint16_t tmp;
         uint16_t operand;
-        int cycles;
+        uint64_t cycles;
         bool halted;
         bool error;
